liquidity_book::core
Module padding

Functions

pad_handle_result

Pad the data and logs in a Response to the block size, with spaces.

pad_query_result

Pad a StdResult<Binary> with spaces

space_pad

Take a Vec and pad it up to a multiple of block_size, using spaces at the end.