liquidity_book::interfaces
Module lb_router

Structs

AddLiquidityResponse

CreateLbPairResponse

GetFactoryResponse

GetIdFromPriceResponse

GetPriceFromIdResponse

GetSwapInResponse

GetSwapOutResponse

ILbRouter

A thin wrapper around ContractInfo that provides additional methods to interact with an LB Router contract.

InstantiateMsg

LiquidityParameters

The liquidity parameters, such as:

Path

The path parameters, such as:

RemoveLiquidityResponse

SwapResponse

Enums

ExecuteMsg

LbRouterError

QueryMsg

Version

This enum represents the version of the pair requested