liquidity_book::interfaces::lb_token
Enum ExecuteMsg

pub enum ExecuteMsg {

    MintTokens {
        mint_tokens: Vec<TokenAmount>,
        memo: Option<String>,
        padding: Option<String>,
    },
    BurnTokens {
        burn_tokens: Vec<TokenAmount>,
        memo: Option<String>,
        padding: Option<String>,
    },
    ChangeMetadata {
        token_id: String,
        public_metadata: Box<Option<Metadata>>,
        private_metadata: Box<Option<Metadata>>,
    },
    Transfer {
        token_id: String,
        from: Addr,
        recipient: Addr,
        amount: Uint256,
        memo: Option<String>,
        padding: Option<String>,
    },
    BatchTransfer {
        actions: Vec<TransferAction>,
        padding: Option<String>,
    },
    Send {
        token_id: String,
        from: Addr,
        recipient: Addr,
        recipient_code_hash: Option<String>,
        amount: Uint256,
        msg: Option<Binary>,
        memo: Option<String>,
        padding: Option<String>,
    },
    BatchSend {
        actions: Vec<SendAction>,
        padding: Option<String>,
    },
    GivePermission {
        allowed_address: Addr,
        token_id: String,
        view_balance: Option<bool>,
        view_balance_expiry: Option<Expiration>,
        view_private_metadata: Option<bool>,
        view_private_metadata_expiry: Option<Expiration>,
        transfer: Option<Uint256>,
        transfer_expiry: Option<Expiration>,
        padding: Option<String>,
    },
    RevokePermission {
        token_id: String,
        owner: Addr,
        allowed_address: Addr,
        padding: Option<String>,
    },
    CreateViewingKey {
        entropy: String,
        padding: Option<String>,
    },
    SetViewingKey {
        key: String,
        padding: Option<String>,
    },
    RevokePermit {
        permit_name: String,
        padding: Option<String>,
    },
    ChangeAdmin {
        new_admin: Addr,
        padding: Option<String>,
    },
    RemoveAdmin {
        current_admin: Addr,
        contract_address: Addr,
        padding: Option<String>,
    },
    RegisterReceive {
        code_hash: String,
        padding: Option<String>,
    },

}

Handle messages to SNIP1155 contract.

Mostly responds with HandleAnswer { <variant_name>: { status: success }} if successful. See HandleAnswer for the response messages for each variant.

Variants

MintTokens

mints additional tokens of existing fungible token_ids, if configuration allows this, ie enable_mint == true. Only minters can access this function

Fields

mint_tokens: Vec<TokenAmount>

memo: Option<String>

padding: Option<String>

BurnTokens

burns existing tokens, if configuration allows this, ie enable_burn == true. Only owners can burn their own tokens in the base specifications. Flexibility is built into the contract functions to allow other addresses to burn tokens, allowed in additional specifications.

Fields

burn_tokens: Vec<TokenAmount>

memo: Option<String>

padding: Option<String>

ChangeMetadata

allows owner or minter to change metadata if allowed by token_id configuration.

Fields

token_id: String

public_metadata: Box<Option<Metadata>>

does not attempt to change if left blank. Can effectively remove metadata by setting metadata to Some(Metadata {token_uri: None, extension: None}) used Box to reduce the total size of the enum variant, to decrease size difference between variants. Not strictly necessary.

private_metadata: Box<Option<Metadata>>

does not attempt to change if left blank. Can effectively remove metadata by setting metadata to Some(Metadata {token_uri: None, extension: None}) used Box to reduce the total size of the enum variant, to decrease size difference between variants. Not strictly necessary.

Transfer

transfers one or more tokens of a single token_id. Other third address can perform this function if it has permission to transfer. ie: if addr3 can call this function to transfer tokens from addr0 to addr2, if addr0 gives addr3 enough transfer allowance.

Fields

token_id: String

from: Addr

recipient: Addr

amount: Uint256

memo: Option<String>

padding: Option<String>

BatchTransfer

performs transfers of multiple token_ids in a single transaction

Fields

actions: Vec<TransferAction>

padding: Option<String>

Send

similar to transfer, but also sends a cosmos message. The recipient needs to be a contract that has a SNIP1155Receive handle function. See receiver for more information.

Fields

token_id: String

from: Addr

recipient: Addr

recipient_code_hash: Option<String>

amount: Uint256

msg: Option<Binary>

memo: Option<String>

padding: Option<String>

BatchSend

performs send of multiple token_ids in a single transaction

Fields

actions: Vec<SendAction>

padding: Option<String>

GivePermission

allows an owner of token_ids to change transfer or viewership permissions to other addresses.

The base specification has three types of permissions:

view balance permission: owner can allow another address to view owner’s balance of specific token_ids

view private metadata: owner can allow another address to view private metadata of specific token_ids

transfer allowance: owner can give permission to another address to transfer tokens up to a certain limit (cumulatively) Owners can set an expiry for each of these permissions.

SNIP1155 gives flexibility for permissions to have any combination of

type of permission granted

on which token_ids

Fields

allowed_address: Addr

address being granted/revoked permission

token_id: String

token id to apply approval/revocation to. Additional Spec feature: if == None, perform action for all owner’s token_ids

view_balance: Option<bool>

optional permission level for viewing balance. If ignored, leaves current permission settings

view_balance_expiry: Option<Expiration>

view_private_metadata: Option<bool>

optional permission level for viewing private metadata. If ignored, leaves current permission settings

view_private_metadata_expiry: Option<Expiration>

transfer: Option<Uint256>

set allowance by for transfer approvals. If ignored, leaves current permission settings

transfer_expiry: Option<Expiration>

padding: Option<String>

optional message length padding

RevokePermission

Removes all permissions that a specific owner has granted to a specific address, for a specific token_id. A permission grantee can use this function to renounce a permission it has been given. For owners, the GivePermission message can be used instead to have the same effect as RevokePermission.

Fields

token_id: String

owner: Addr

token owner

allowed_address: Addr

address which has permission

padding: Option<String>

CreateViewingKey

Fields

entropy: String

padding: Option<String>

SetViewingKey

Fields

key: String

padding: Option<String>

RevokePermit

disallow the use of a query permit

Fields

permit_name: String

padding: Option<String>

ChangeAdmin

Fields

new_admin: Addr

padding: Option<String>

RemoveAdmin

Permanently breaks admin keys for this contract. No admin function can be called after this action. Any existing curators or minters will remain as curators or minters; no new curators can be added and no current curator can be removed.

Requires caller to input current admin address and contract address. These inputs are not strictly necessary, but as a safety precaution to reduce the chances of accidentally calling this function.

Fields

current_admin: Addr

contract_address: Addr

padding: Option<String>

RegisterReceive

Fields

code_hash: String

padding: Option<String>

Implementations

impl ExecuteMsg

pub fn to_cosmos_msg( &self, code_hash: String, contract_addr: String, send_amount: Option<Uint128>, ) -> StdResult<CosmosMsg>