liquidity_book::libraries::bin_helper
Trait BinHelper

pub trait BinHelper {
    // Required methods
    fn get_amount_out_of_bin(
        &self,
        amount_to_burn: U256,
        total_supply: U256,
    ) -> Result<Bytes32, BinError>;
    fn get_shares_and_effective_amounts_in(
        &self,
        amounts_in: Bytes32,
        price: U256,
        total_supply: U256,
    ) -> Result<(U256, Bytes32), BinError>;
    fn get_liquidity(&self, price: U256) -> Result<U256, BinError>;
    fn verify_amounts(&self, active_id: u32, id: u32) -> Result<(), BinError>;
    fn get_composition_fees(
        &self,
        parameters: PairParameters,
        bin_step: u16,
        amounts_in: Bytes32,
        total_supply: U256,
        shares: U256,
    ) -> Result<Bytes32, BinError>;
    fn is_empty(&self, is_x: bool) -> bool;
    fn get_amounts(
        &self,
        parameters: PairParameters,
        bin_step: u16,
        swap_for_y: bool,
        active_id: u32,
        amounts_in_left: Bytes32,
    ) -> Result<(Bytes32, Bytes32, Bytes32), BinError>;
    fn received(&self, token_x: u128, token_y: u128) -> Bytes32;
    fn received_x(&self, token_x: u128) -> Bytes32;
    fn received_y(&self, token_y: u128) -> Bytes32;
}

Required Methods

fn get_amount_out_of_bin( &self, amount_to_burn: U256, total_supply: U256, ) -> Result<Bytes32, BinError>

Returns the amount of tokens that will be received when burning the given amount of liquidity.

fn get_shares_and_effective_amounts_in( &self, amounts_in: Bytes32, price: U256, total_supply: U256, ) -> Result<(U256, Bytes32), BinError>

Returns the share and the effective amounts in when adding liquidity.

fn get_liquidity(&self, price: U256) -> Result<U256, BinError>

Returns the amount of liquidity following the constant sum formula L = price * x + y.

fn verify_amounts(&self, active_id: u32, id: u32) -> Result<(), BinError>

Verify that the amounts are correct and that the composition factor is not flawed.

fn get_composition_fees( &self, parameters: PairParameters, bin_step: u16, amounts_in: Bytes32, total_supply: U256, shares: U256, ) -> Result<Bytes32, BinError>

Returns the composition fees when adding liquidity to the active bin with a different composition factor than the bin’s one, as it does an implicit swap.

fn is_empty(&self, is_x: bool) -> bool

Returns whether the bin is empty (true) or not (false).

fn get_amounts( &self, parameters: PairParameters, bin_step: u16, swap_for_y: bool, active_id: u32, amounts_in_left: Bytes32, ) -> Result<(Bytes32, Bytes32, Bytes32), BinError>

Returns the amounts of tokens that will be added and removed from the bin during a swap along with the fees that will be charged.

fn received(&self, token_x: u128, token_y: u128) -> Bytes32

Returns the encoded amounts that were transferred to the contract for both tokens.

fn received_x(&self, token_x: u128) -> Bytes32

Returns the encoded amounts that were transferred to the contract, only for token X.

fn received_y(&self, token_y: u128) -> Bytes32

Returns the encoded amounts that were transferred to the contract, only for token Y.