liquidity_book::libraries::fee_helper
Trait FeeHelper

pub trait FeeHelper {
    // Required methods
    fn verify_fee(&self) -> Result<(), FeeError>;
    fn verify_protocol_share(&self) -> Result<(), FeeError>;
    fn get_fee_amount_from(&self, total_fee: u128) -> Result<u128, FeeError>;
    fn get_fee_amount(&self, total_fee: u128) -> Result<u128, FeeError>;
    fn get_composition_fee(&self, total_fee: u128) -> Result<u128, FeeError>;
    fn get_protocol_fee_amount(
        &self,
        protocol_share: u128,
    ) -> Result<u128, FeeError>;
}

Trait for fee-related operations

Required Methods

fn verify_fee(&self) -> Result<(), FeeError>

Check that the fee is not too large.

fn verify_protocol_share(&self) -> Result<(), FeeError>

Check that the protocol share is not too large.

fn get_fee_amount_from(&self, total_fee: u128) -> Result<u128, FeeError>

Calculates the fee amount from the amount with fees, rounding up.

fn get_fee_amount(&self, total_fee: u128) -> Result<u128, FeeError>

Calculates the fee amount that will be charged, rounding up.

fn get_composition_fee(&self, total_fee: u128) -> Result<u128, FeeError>

Calculates the composition fee amount from the amount with fees, rounding down.

fn get_protocol_fee_amount( &self, protocol_share: u128, ) -> Result<u128, FeeError>

Calculates the protocol fee amount from the fee amount and the protocol share, rounding down.