liquidity_book::libraries
Module hooks

Structs

HooksParameters

Parameters

Constants

AFTER_BURN

AFTER_BURN_FLAG

AFTER_FLASH_LOAN

AFTER_FLASH_LOAN_FLAG

AFTER_MINT

AFTER_MINT_FLAG

AFTER_SWAP

AFTER_SWAP_FLAG

AFTER_TRANSFER

AFTER_TRANSFER_FLAG

BEFORE_BURN

BEFORE_BURN_FLAG

BEFORE_FLASH_LOAN

BEFORE_FLASH_LOAN_FLAG

BEFORE_MINT

BEFORE_MINT_FLAG

BEFORE_SWAP

BEFORE_SWAP_FLAG

BEFORE_TRANSFER

BEFORE_TRANSFER_FLAG

Functions

after_swap

@dev Helper function to call the afterSwap function on the hooks contract, only if the AFTER_SWAP_FLAG is set in the hooksParameters @param hooksParameters The encoded hooks parameters @param sender The sender @param to The recipient @param swapForY Whether the swap is for Y @param amountsOut The amounts out

after_transfer_from

Helper function to call the afterTransferFrom function on the hooks contract, only if the AFTER_TRANSFER_FLAG is set in the hooksParameters

before_swap

@dev Helper function to call the beforeSwap function on the hooks contract, only if the BEFORE_SWAP_FLAG is set in the hooksParameters @param hooksParameters The encoded hooks parameters @param sender The sender @param to The recipient @param swapForY Whether the swap is for Y @param amountsIn The amounts in

before_transfer_from

Helper function to call the beforeTransferFrom function on the hooks contract, only if the BEFORE_TRANSFER_FLAG is set in the hooksParameters

code_hash_to_bytes32

Converts a hex-encoded code hash (SHA-256) into a Bytes32 array. The input should be a 64-character hex string.

decode

@dev Helper function to decode the hooks parameters from a single bytes32 value @param hooksParameters The encoded hooks parameters @return parameters The hooks parameters

encode

@dev Helper function to encode the hooks parameters to a single bytes32 value @param parameters The hooks parameters @return hooksParameters The encoded hooks parameters

get_flags

@dev Helper function to get the flags from the encoded hooks parameters @param hooksParameters The encoded hooks parameters @return flags The flags

get_hooks

@dev Helper function to get the hooks address from the encoded hooks parameters @param hooksParameters The encoded hooks parameters @return hooks The hooks address

on_hooks_set

@dev Helper function call the onHooksSet function on the hooks contract, only if the hooksParameters is not 0 @param hooksParameters The encoded hooks parameters @param onHooksSetData The data to pass to the onHooksSet function

set_hooks

@dev Helper function to set the hooks address in the encoded hooks parameters @param hooksParameters The encoded hooks parameters @param newHooks The new hooks address @return hooksParameters The updated hooks parameters