liquidity_book::libraries::math
Module packed_u128_math

Liquidity Book Tree Math Library

Author: Kent and Haseeb

This module contains functions to encode and decode two u128 into a single Bytes32 and interact with the encoded Bytes32.

u128 is a 128-bit unsigned integer type, which means that its little-endian byte representation is 16 bytes long. A Bytes32 value is a [u8; 32] and can hold 256 bits, or two u128 values.

Enums

PackedUint128MathError

Constants

BASIS_POINT_MAX

Traits

PackedUint128Math