liquidity_book::libraries::math::packed_u128_math
Trait PackedUint128Math

pub trait PackedUint128Math: From<[u8; 32]> + AsRef<[u8]> {

    // Provided methods
    fn min() -> Self { ... }
    fn max() -> Self { ... }
    fn encode(x1: u128, x2: u128) -> Self { ... }
    fn encode_alt(x: u128, first: bool) -> Self { ... }
    fn encode_first(x1: u128) -> Bytes32 { ... }
    fn encode_second(x2: u128) -> Bytes32 { ... }
    fn decode(&self) -> (u128, u128) { ... }
    fn decode_alt(&self, first: bool) -> u128 { ... }
    fn decode_x(&self) -> u128 { ... }
    fn decode_y(&self) -> u128 { ... }
    fn add(&self, y: Bytes32) -> Result<Self, PackedUint128MathError> { ... }
    fn add_alt(
        &self,
        y1: u128,
        y2: u128,
    ) -> Result<Self, PackedUint128MathError> { ... }
    fn sub(&self, y: Self) -> Result<Self, PackedUint128MathError> { ... }
    fn sub_alt(
        &self,
        y1: u128,
        y2: u128,
    ) -> Result<Self, PackedUint128MathError> { ... }
    fn lt(&self, y: Self) -> bool { ... }
    fn gt(&self, y: Self) -> bool { ... }
    fn scalar_mul_div_basis_point_round_down(
        &self,
        multiplier: u128,
    ) -> Result<Self, Error> { ... }

}

Provided Methods

fn min() -> Self

fn max() -> Self

fn encode(x1: u128, x2: u128) -> Self

Encodes two u128 values into a single Bytes32 value, with the first u128 value stored in the first 128 bits of the Bytes32 value and the second u128 value stored in the last 128 bits.

Arguments

x1 - The first u128 value to encode.

x2 - The second u128 value to encode.

Returns

The encoded Bytes32 value, with x1 stored in the first 128 bits and x2 stored in the last 128 bits.

fn encode_alt(x: u128, first: bool) -> Self

Encodes a u128 value into a single Bytes32 value, with the u128 value stored in either the first or last 128 bits of the Bytes32 value. The remaining 128 bits are set to zero.

Arguments

x - The u128 value to encode.

first - Whether to encode as the first or second u128.

Returns

The encoded Bytes32 value, with the u128 value stored in either the first or last 128 bits of the Bytes32 value. The remaining 128 bits are set to zero.

fn encode_first(x1: u128) -> Bytes32

Encodes a u128 value into a single Bytes32 value, with the u128 value stored in the first 128 bits of the Bytes32 value and the last 128 bits set to zero.

Arguments

x1 - The u128 value to encode.

Returns

The encoded Bytes32 value, with x1 stored in the first 128 bits and the last 128 bits set to zero.

fn encode_second(x2: u128) -> Bytes32

Encodes a u128 value into a single Bytes32 value, with the u128 value stored in the last 128 bits of the Bytes32 value and the first 128 bits set to zero.

Arguments

x2 - The u128 value to encode.

Returns

The encoded Bytes32 value, with x2 stored in the last 128 bits and the first 128 bits set to zero.

fn decode(&self) -> (u128, u128)

Decodes a Bytes32 value into two u128 values as follows:

[0 - 128[: x1

[128 - 256[: x2

Returns

A tuple of two u128 values representing the decoded x1 and x2 values.

fn decode_alt(&self, first: bool) -> u128

Decodes a Bytes32 value into a u128 value as the first or second u128.

Arguments

first - A boolean value indicating whether to decode the first chunk (true) or the second chunk (false)

Returns

A u128 value representing the decoded x1 or x2 value, depending on the value of first.

fn decode_x(&self) -> u128

Decodes a Bytes32 value into a u128 value as the first chunk.

Returns

A u128 value representing the decoded x1 value.

fn decode_y(&self) -> u128

Decodes a Bytes32 value into a u128 value as the second chunk.

Returns

A u128 value representing the decoded x2 value.

fn add(&self, y: Bytes32) -> Result<Self, PackedUint128MathError>

Adds two Bytes32 values encoded as follows:

[0 - 128[: x1

[128 - 256[: x2

[0 - 128[: y1

[128 - 256[: y2

Arguments

x - The first Bytes32 value represented as a [u8; 32] array

y - The second Bytes32 value represented as a [u8; 32] array

Returns

A Bytes32 value represented as a [u8; 32] array, encoding the sum of x and y as follows:

[0 - 128[: x1 + y1

[128 - 256[: x2 + y2

fn add_alt(&self, y1: u128, y2: u128) -> Result<Self, PackedUint128MathError>

Adds a Bytes32 value encoded as follows:

[0 - 128[: x1

[128 - 256[: x2

and two u128 values y1 and y2.

Arguments

x - The Bytes32 value represented as a [u8; 32] array

y1 - The first u128 value

y2 - The second u128 value

Returns

A Bytes32 value represented as a [u8; 32] array, encoding the sum of x and (y1, y2) as follows:

[0 - 128[: x1 + y1

[128 - 256[: x2 + y2

fn sub(&self, y: Self) -> Result<Self, PackedUint128MathError>

Subtracts two Bytes32 values encoded as follows:

[0 - 128[: x1

[128 - 256[: x2

[0 - 128[: y1

[128 - 256[: y2

Arguments

x - The first Bytes32 value represented as a [u8; 32] array

y - The second Bytes32 value represented as a [u8; 32] array

Returns

A Bytes32 value represented as a [u8; 32] array, encoding the difference between x and y as follows:

[0 - 128[: x1 - y1

[128 - 256[: x2 - y2

fn sub_alt(&self, y1: u128, y2: u128) -> Result<Self, PackedUint128MathError>

Subtracts a Bytes32 value encoded as follows:

[0 - 128[: x1

[128 - 256[: x2

and two u128 values y1 and y2.

Arguments

x - The Bytes32 value represented as a [u8; 32] array

y1 - The first u128 value

y2 - The second u128 value

Returns

A Bytes32 value represented as a [u8; 32] array, encoding the difference between x and (y1, y2) as follows:

[0 - 128[: x1 - y1

[128 - 256[: x2 - y2

fn lt(&self, y: Self) -> bool

Returns whether any of the u128 of x is strictly less than the corresponding u128 of y.

x and y are both encoded as Bytes32 with the following structure:

[0-15]: first u128 value

[16-31]: second u128 value

Arguments

x - A Bytes32 encoding a pair of u128 values.

y - A Bytes32 encoding a pair of u128 values.

Returns

true if x is less than y, and false otherwise.

fn gt(&self, y: Self) -> bool

Returns whether any of the u128 of x is strictly greater than the corresponding u128 of y.

x and y are both encoded as Bytes32 with the following structure:

[0-15]: first u128 value

[16-31]: second u128 value

Arguments

x - A Bytes32 encoding a pair of u128 values.

y - A Bytes32 encoding a pair of u128 values.

Returns

true if x is greater than y, and false otherwise.

fn scalar_mul_div_basis_point_round_down( &self, multiplier: u128, ) -> Result<Self, Error>

Multiplies an encoded Bytes32 by a u128 then divides the result by 10_000, rounding down.

The result can’t overflow as the multiplier needs to be smaller or equal to 10_000.

Arguments

x - The Bytes32 encoded as follows:

[0 - 128[ : x1

[128 - 256[ : x2

multiplier - The u128 to multiply by (must be smaller or equal to 10_000).

Returns

Returns the product of x and multiplier encoded as follows:

[0 - 128[ : floor((x1 * multiplier) / 10_000)

[128 - 256[ : floor((x2 * multiplier) / 10_000)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.