liquidity_book::libraries::math
Module u256x256_math

Liquidity Book U256x256 Math Library

Author: Kent and Haseeb

Helper library used for full precision calculations.

Enums

U256x256MathError

Traits

U256ToU512Conversion

U256x256Math

U512ToU256Conversion

Functions

_get_end_of_div_round_down

Helper function to return the result of x * y / denominator with full precision

_get_mul_prods 🔒

Helper function to return the result of x * y as 2 U256

addmod

Computes (x + y) % k where the addition is performed with arbitrary precision and does not wrap around at 2^256.

mulmod 🔒

Computes (x * y) % k where the multiplication is performed with arbitrary precision and does not wrap around at 2^256.