liquidity_book::libraries::pair_parameter_helper
Struct PairParameters

pub struct PairParameters(pub Bytes32);

Tuple Fields
0: Bytes32
Implementations

impl PairParameters

pub fn get_base_factor(&self) -> u16

Get the base factor from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 16[: base factor (16 bits)

[16 - 256[: other parameters

pub fn get_filter_period(&self) -> u16

Get the filter period from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 16[: other parameters

[16 - 28[: filter period (12 bits)

[28 - 256[: other parameters

pub fn get_decay_period(&self) -> u16

Get the decay period from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 16[: other parameters

[28 - 40[: decay period (12 bits)

[40 - 256[: other parameters

pub fn get_reduction_factor(&self) -> u16

Get the reduction factor from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 40[: other parameters

[40 - 54[: reduction factor (14 bits)

[54 - 256[: other parameters

pub fn get_variable_fee_control(&self) -> u32

Get the variable fee control from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 54[: other parameters

[54 - 78[: variable fee control (24 bits)

[78 - 256[: other parameters

pub fn get_protocol_share(&self) -> u16

Get the protocol share from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 78[: other parameters

[78 - 92[: protocol share (14 bits)

[92 - 256[: other parameters

pub fn get_max_volatility_accumulator(&self) -> u32

Get the max volatility accumulator from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 92[: other parameters

[92 - 112[: max volatility accumulator (20 bits)

[112 - 256[: other parameters

pub fn get_volatility_accumulator(&self) -> u32

Get the volatility accumulator from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 112[: other parameters

[112 - 132[: volatility accumulator (20 bits)

[132 - 256[: other parameters

pub fn get_volatility_reference(&self) -> u32

Get the volatility reference from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 132[: other parameters

[132 - 152[: volatility reference (20 bits)

[152 - 256[: other parameters

pub fn get_id_reference(&self) -> u32

Get the index reference from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 152[: other parameters

[152 - 176[: index reference (24 bits)

[176 - 256[: other parameters

pub fn get_time_of_last_update(&self) -> u64

Get the time of last update from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 176[: other parameters

[176 - 216[: time of last update (40 bits)

[216 - 256[: other parameters

pub fn get_oracle_id(&self) -> u16

Get the oracle id from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 216[: other parameters

[216 - 232[: oracle id (16 bits)

[232 - 256[: other parameters

pub fn get_active_id(&self) -> u32

Get the active index from the encoded pair parameters.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 232[: other parameters

[232 - 256[: active index (24 bits)

pub fn get_delta_id(&self, active_id: u32) -> u32

Get the delta between the current active index and the cached active index.

Arguments

parameters - The encoded pair parameters, as follows:

[0 - 232[: other parameters

[232 - 256[: active index (24 bits)

active_id - The current active index

pub fn get_base_fee(&self, bin_step: u16) -> u128

Calculates the base fee, with 18 decimals.

Arguments

parameters - The encoded pair parameters

bin_step - The bin step (in basis points)

pub fn get_variable_fee(&self, bin_step: u16) -> u128

Calculates the variable fee.

Arguments

parameters - The encoded pair parameters

bin_step - The bin step (in basis points)

pub fn get_total_fee(&self, bin_step: u16) -> Result<u128, PairParametersError>

Calculates the total fee, which is the sum of the base fee and the variable fee.

Arguments

parameters - The encoded pair parameters

bin_step - The bin step (in basis points)

pub fn set_oracle_id(&mut self, oracle_id: u16) -> &mut Self

Set the oracle id in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

oracle_id - The oracle id

pub fn set_volatility_reference( &mut self, vol_ref: u32, ) -> Result<&mut Self, PairParametersError>

Set the volatility reference in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

vol_ref - The volatility reference

pub fn set_volatility_accumulator( &mut self, vol_acc: u32, ) -> Result<&mut Self, PairParametersError>

Set the volatility accumulator in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

vol_acc - The volatility accumulator

pub fn set_active_id( &mut self, active_id: u32, ) -> Result<&mut Self, PairParametersError>

Set the active id in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

active_id - The active id

pub fn set_static_fee_parameters( &mut self, base_factor: u16, filter_period: u16, decay_period: u16, reduction_factor: u16, variable_fee_control: u32, protocol_share: u16, max_volatility_accumulator: u32, ) -> Result<&mut Self, PairParametersError>

Sets the static fee parameters in the encoded pair parameters.

Arguments

base_factor - The base factor

filter_period - The filter period

decay_period - The decay period

reduction_factor - The reduction factor

variable_fee_control - The variable fee control

protocol_share - The protocol share

max_volatility_accumulator - The max volatility accumulator

pub fn update_id_reference(&mut self) -> &mut Self

Updates the index reference in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

pub fn update_time_of_last_update( &mut self, time: u64, ) -> Result<&mut Self, PairParametersError>

Updates the time of last update in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

current_time - The current timestamp

pub fn update_volatility_reference( &mut self, ) -> Result<&mut Self, PairParametersError>

Updates the volatility reference in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

pub fn update_volatility_accumulator( &mut self, active_id: u32, ) -> Result<&mut Self, PairParametersError>

Updates the volatility accumulator in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

active_id - The active id

pub fn update_references( &mut self, time: u64, ) -> Result<&mut Self, PairParametersError>

Updates the volatility reference and the volatility accumulator in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

pub fn update_volatility_parameters( &mut self, active_id: u32, time: u64, ) -> Result<&mut Self, PairParametersError>

Updates the volatility reference and the volatility accumulator in the encoded pair parameters.

Arguments

parameters - The encoded pair parameters

active_id - The active id